* history.c (read_hrecs): st_blksize is unsigned long, not int. 	This isn't just cosmetic - getting it wrong will cause coredumps 	and such on 64 bit machines.
CVS-ID: src/ChangeLog 1.1838
CVS-ID: src/history.c 1.47

+++ src/history.c
@@ -1075,8 +1075,8 @@ read_hrecs (fname)
 	{
 	    if (nl - cp >= STAT_BLOCKSIZE(st_buf))
 	    {
-		error(1, 0, "history line too long (> %d)",
-		      STAT_BLOCKSIZE(st_buf));
+		error(1, 0, "history line too long (> %lu)",
+		      (unsigned long) STAT_BLOCKSIZE(st_buf));
 	    }
 	    if (nl > cp)
 		memmove (cpstart, cp, nl - cp);
